Finished assignment + Read Desc
Two things about this. The "notsidious.ogg" in this Godot project gets cut off when the project gets played; probably because of how long it is or maybe it's just too large, but previewing the sound in the editing mode/outside of game mode works.

The music has a large gap of silence in the beginnning; I couldn't slim this down, or at least I couldn't in Musescore. There's a song there, I promise, and it's adequate, not some jumble of notes.